Don't encode scrypt parameters in password hash
The current hashing code is designed for the case where
the server stores the hash in secure location, and
validates an unknown password against it. IOW the hash
data is trustworthy at this time. It is not obvious
from the API, however, that the hash must be treated
as secure data, so its a bad idea to store the scrypt
params in the hash. Always use the built-in parameters
when validating passwords against the hash so we have
better protection against tampering.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>